Clean up imports and modules
