+++ src/test/mod.rs
@@ -33,9 +33,6 @@ use std::sync::{Arc, Mutex};
 use std::time::Instant;
 use url::Url;
 
-#[path = "../../tests/support/mod.rs"]
-mod support;
-
 fn initialize(id: usize, root_path: Option<String>) -> Request<ls_server::InitializeRequest> {
     initialize_with_opts(id, root_path, None)
 }
